Add fixes for details and index
I needed to change the sort of the properties in detail.scss, the import of the fonts in the index, use the universal selector to not redefine the font-family property and reagroup some classes that have the same styles.
Also I needed to create the _colors.scss to applicate it with the styles.