Add responsive for the book details
I added the responsive feature for adapt the book details for screens under 1024px, for this I had to change some margins, some heights and widhts for the logo and the badge. Also, the badge needed to be 30% smaller than the original. So I made use of the CSS variables, and made that with the responsive, the badge image only is going to be a 70% of the original badge. (70/100)*total = 70% of the total.