Add Angular CLI structure with register html
In this feature we are creating the Angular CLI and we separate the 2 trainings in two different folders. maquetado and wolox-books. The folder wolox-books is the folder that have all the angular training.
In this commit Im creating the Angular CLI in the folder, and also making the html, and scss move from maquetado to this angular step.